storage: Rework data structure for extra handlers.
Use a dictionary: { handler_type : config }

A store handler's only unique identifier is the pattern of model types
it handles.  That means for "extra" store handlers (those not associated
with any model types), there can only be one instance per handler type.

This may not be a problem in practice, but it's a theoretical constraint
to the internal data structures.

Also note, these "extra" store handlers are never actually instantiated.
But that will change shortly...