storage: Honor Host.source attribute.
This now actually instantiates "extra" store handlers.

A subtle, and perhaps unexpected, constraint with this is the "source"
attribute can only refer to a store handler with no models associated
with it.  Otherwise the "source" value alone would not necessarily
uniquely identify the desired store handler (see previous commit).

If this becomes an issue, we may need to give store handlers a "name"
attribute or some other explicitly unique identifier.