storage: Give storage handlers a unique name.
To better address the uniqueness issue discussed in the previous two
commit messages, this commit breaks backward compatibility in storage
handler definitions in a configuration file.

The module name is now specified in a "type" member, rather than a
"name" member.  The "name" member is now the unique identifier, but
specifying it in a storage handler definition is optional.  If left
unspecified, it inherits the module name from the "type" member.
If that conflicts with an existing definition, an integer suffix is
tried until it arrives at a unique name.

So if multiple storage handlers of type "commissaire.storage.etcd"
all lack an explicit "name", their generated names would be:

    "commissaire.storage.etcd"
    "commissaire.storage.etcd-1"
    "commissaire.storage.etcd-2"
    ...

The Host.source attribute now refers to these storage handler names.